Get rid of the static kernel configuration files
With a backported patch from the OpenWrt trunk, it is now easy to generate an
equivalent configuration using CONFIG_ALL_KMODS.

The build will take a bit longer because all kernel module packages are actually
built even when they are not included in the image, but adding new targets
becomes a lot easier.

Also, related documentation updates and fixes.