gluon-mesh-batman-adv-core: Always forward towards mesh/batman-adv
Since we just segmented the IGMP/MLD domain between node and
mesh boundaries, we will not be able to keep track of multicast
listeners on other nodes from the bridge anymore.

Therefore we need to always hand any multicast packets we received from
local clients to batman-adv. We can do this by setting the bridge port
option "multicast_router" to 2.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>