Add support for TP-Link TL-WA901ND-v4
I veriefied the functionality of this device, IBSS and 802.11s meshing, Ethernet, LEDs and Buttons, WiFi works perfectly with this device. The primary MAC is the same as printed on the devices sticker.

The firmware cannot be flashed via the stock firmwares web ui (OpenWrt and LEDE can't be flashed either). The firmware has to be flashed onto the device via the TFTP method mentioned in [this post](https://forum.openwrt.org/viewtopic.php?pid=320942#p320942).

In short: Download the firmware and rename it to wa901ndv4_tp_recovery.bin, setup a tftp server and configure a static IP address (192.168.0.66), press (and hold for a few seconds) the reset button while turning the device on. The device is now downloading the firmware and you should be able to configure your new Gluon node after the device has rebooted.