change fastd sysloglevel to warn instead of verbose
it is easy to set this to a higher value again if needed. This will prevent fastd from flooding the log ringbuffer, even when it is not needed/no uplink via fastd. 
This flooding raise with every peer known from site.conf.
this solve partly https://github.com/freifunk-gluon/gluon/issues/896 - while it is still not configurable