kernel: Add uapi/batman_adv.h from batman-adv 2016.4
The batman_adv.h is required by userspace applications to send/receive
generic netlink messages. This file either has to be included in each
program which uses it (like iw/hostapd/wavemon/... does it for nl80211.h)
or the included file in the uapi kernel headers has to be patched to always
be at the correct version.

gluon decided to prefer a patched kernel to reduce the number of copies of
this file in its own sources.

Signed-off-by: Sven Eckelmann <sven@narfation.org>