ar71xx-generic: don't try to build a factory image for D-Link DIR-825 B1
The size of the factory images is limited to 4MB, which caused build
failures when many additional packages were included.

Rather than moving the device to ar71xx-tiny, we ignore the factory images
and just build the sysupgrade. This way, the whole flash is usable for
Gluon.

This means that installing Gluon on these devices will now require to flash
a plain LEDE image first, and then upgrading using the Gluon sysupgrade
image.

Fixes #1020