gluon-mesh-batman-adv: unconditionally rate limit ARP
With a reasoning similar to "gluon-ebtables: unconditionally segment
IGMP/MLD" also make the ARP rate limiting mandatory.

It turned out to be very common that there is a client device with
an application scanning the IP subnet, causing congestion and high
load for any community which did not add the gluon-ebtables-limit-arp
package yet.

Therefore just always add gluon-ebtables-limit-arp via a dependency.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>