build: add support for other images
In light of future devices in need of special images for initial
flashing which are, in contrast to factory images, only flashable
from the bootloader, we add a new category for images which don't fit
into factory nor sysupgrade.

Devices in need for those images contain the AVM Fritz!Box 4040 and
various TP-Link Mediatek models.

Example for the Fritz!Box 4040:

 > other -squashfs-eva.bin bootloader.bin

results in the eva image landing in the "other" folder suffixed with
"bootlaoder.bin".