build: add support for bootloader images
In light of future devices in need of special images for initial
flashing which are, in contrast to factory images, only flashable
from the bootloader, we add a new category for images which are only
suitable for flashing thru the vendor bootloader.

Devices in need for those images contain the AVM Fritz!Box 4040 and
various TP-Link Mediatek models.