ar71xx: ag71xx: reduce NAPI weight and default size of TX/RX ring buffers
As to https://lwn.net/Articles/139884/ a too large NAPI weight can
cause increased system load as a driver may turn off polling and
return to interrupt-driven mode if the quota was not filled in
the last polling cycle creating the need of more interrupts to 
be handled. This commit decreases the NAPI weight from 64 to 32.

Furthermore this commit decreases the default TX/RX ring buffer
sizes from 128/256 to 64/128 as the buffers will unlikely be 
filled with a throughput <= 100 Mbit/s.