ar71xx: add support for TP-Link RE355
Support is marked as broken because this device is sold in two variants,
one with 64M RAM and another with 128M. As of now 64M is not enough
for ath10k with 5GHz enabled.
As there is no indication known which variant one will get better mark
it as broken.

Signed-off-by: Christoph Krapp <achterin@googlemail.com>