ar71xx: flag FritzBox 4020 buttons as active low
Buttons of AVM FritzBox 4020 are incorrectly flagged as active high.

This was an oversight as RFKill button was working as expected even
with incorrectly flagged GPIO.

Config mode is still accessable but will start on any button press, not
only for the ones >= 10 seconds.

This commit corrects this behavior to the expected one (10 second long
press necessary to enter config mode).