gluon-mesh-batman-adv: Add IPv4 subnet to br-client
This allows us to route and NAT between br-client and br-wan in the future.
Which is necessary for the gluon-alt-esc-provider package, for instance.

With the local-node routes separated in their own table, the issue
described in:

3ef28a46842b ("gluon-client-bridge: Revert "move IPv4 local subnet route to br-client (#1312)"

and initially introduced in:

b3762fc61cca ("gluon-client-bridge: move IPv4 local subnet route to br-client (#1312)")

does not apply anymore, which makes this change safe again.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>