gluon-web-model: Add support for "Supramaps"
It was not possible to modularize single sections of forms in the config mode.
This commit adds support for ´Supramaps´. Using supramaps it is possible to create an additional directory where inframaps reside.
E.g. a Lua model file with a same-named directory (excluding the file extension) defines a Form element and returns it. This is a supramap.
If Lua files exist in that directoryn the Form element is available in their environment as "Supramap".
Thus these files can add additional sections which are "inframaps" and must return the modified supramap.
This allows e.g. modularizing the password field in the gluon-web-admin package as an own package without hampering usability.