gluon-mesh-batman-adv-14: remove CONFLICTs line with v15
batman-adv-legacy now uses its own namespaces. And whether to use v14 or
v15 is determined by the routing algorithm specified in the site
configuration. Which means both the batman-adv and batman-adv-legacy
kernel modules as well as the according glue code provided by
gluon-mesh-batman-adv-14 and gluon-mesh-batman-adv-15 can be installed
at the same time.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>