gluon-core: run firewall upgrade script after basic network setup
Reorder scripts so that the mesh_lan interface is accounted for.

Two other firewall upgrade scripts (mesh-babel and l3roamd) are
reordered as well. While there seems to be no hard dependency at the
moment, it makes sense to run the basic setup first, also to avoid
problems with future changes.

Closes: #2090
Fixes: ed094bc68c7c ("gluon-core: firewall: Allow custom gluon_wired interfaces (#2041)")