openwrt base-files: seperate generation functions
Move the functions responsible for generating the initial UCI board
configuration to a seperate sourcable file.

This way, downstream projects (such as Gluon) can take advantage of them
for allowing a migration from ar71xx to ath79, as the configuration
options can be abstracted and re-created using a fresh board
configuration.