kernel: mtd: add support for EN25QH64 in spi-nor.c
The Eon EN25QH64 is a 64 Mbit SPI NOR flash memory chip. Its 32, 128 and
256 Mbits siblings are supported upstream but this particular size
wasn't.
This commit includes patches for kernels 4.14 and 4.19.

Tested on a COMFAST CF-E120A v3 (ath79).

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>

---
This patch was taken from upstream OpenWRT commit-id
359f5e539036db4f7ac69a6d1c3fb7fe70266ffd.

Additionally change needed for CPE210 V3.2 was backporte from 4.14 to
4.9.
This allows us to use The CPE 210 V3.2 with Gluon 2019.x.